+++ src/input/Mouse.js
@@ -172,22 +172,6 @@ Phaser.Mouse.BACK_BUTTON = 3;
 */
 Phaser.Mouse.FORWARD_BUTTON = 4;
 
-/**
- * @constant
- * @type {number}
- * @deprecated
- * @see Phaser.MouseWheel.UP
- */
-Phaser.Mouse.WHEEL_UP = 1;
-
-/**
- * @constant
- * @type {number}
- * @deprecated
- * @see Phaser.MouseWheel.DOWN
- */
-Phaser.Mouse.WHEEL_DOWN = -1;
-
 Phaser.Mouse.prototype = {
 
     /**
@@ -497,38 +481,6 @@ Phaser.Mouse.prototype = {
     },
 
     /**
-    * Exit a pointer-locked state.
-    * @method Phaser.Mouse#releasePointerLock
-    * @deprecated
-    * @see Phaser.PointerLock#exit
-    */
-    releasePointerLock: function ()
-    {
-
-        console.warn('Deprecated. Please use `input.pointerLock.exit()` instead.');
-
-        return this.input.pointerLock.exit();
-
-    },
-
-    /**
-    * If the browser supports it you can request that the pointer be locked to the browser window.
-    * This is classically known as 'FPS controls', where the pointer can't leave the browser until the user presses an exit key.
-    * If the browser successfully enters a locked state the event Phaser.Mouse.pointerLock will be dispatched and the first parameter will be 'true'.
-    * @method Phaser.Mouse#requestPointerLock
-    * @deprecated
-    * @see Phaser.PointerLock#request
-    */
-    requestPointerLock: function ()
-    {
-
-        console.warn('Deprecated. Please use `input.pointerLock.request()` instead.');
-
-        return this.input.pointerLock.request();
-
-    },
-
-    /**
     * Stop the event listeners.
     * @method Phaser.Mouse#stop
     */
@@ -553,71 +505,3 @@ Phaser.Mouse.prototype = {
 };
 
 Phaser.Mouse.prototype.constructor = Phaser.Mouse;
-
-/**
-* If the mouse has been Pointer Locked successfully this will be set to true.
-*
-* @name Phaser.Mouse#locked
-* @property {boolean} locked
-* @default
-* @deprecated
-* @see Phaser.PointerLock#locked
-*/
-Object.defineProperty(Phaser.Mouse.prototype, 'locked', {
-
-    get: function ()
-    {
-        return this.input.pointerLock.locked;
-    }
-
-});
-
-/**
-* This event is dispatched when the browser enters or leaves pointer lock state.
-*
-* @name Phaser.Mouse#pointerLock
-* @property {Phaser.Signal} pointerLock
-* @default
-* @deprecated
-* @see Phaser.PointerLock#onChange
-*/
-Object.defineProperty(Phaser.Mouse.prototype, 'pointerLock', {
-
-    get: function ()
-    {
-        return this.input.pointerLock.onChange;
-    }
-
-});
-
-/**
- * @property {function} mouseWheelCallback - A callback that can be fired when the mousewheel is used.
- * @deprecated
- * @see Phaser.MouseWheel#callback
- */
-Object.defineProperty(Phaser.Mouse.prototype, 'mouseWheelCallback', {
-
-    get: function ()
-    {
-        return this.input.mouseWheel.callback;
-    },
-
-    set: function (val)
-    {
-        this.input.mouseWheel.callback = val;
-    }
-
-});
-
-/**
- * The direction of the _last_ mousewheel usage. 1 for up; -1 for down.
- * @property {number} wheelDelta
- * @deprecated
- * @see Phaser.MouseWheel#delta
- */
-Object.defineProperty(Phaser.Mouse.prototype, 'wheelDelta', {
-    get: function ()
-    {
-        return this.input.mouseWheel.delta;
-    }
-});